requirements.txt
# Core LLM & Agent Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.2.0

# PDF Extraction
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
Pillow>=10.0.0

# Vector Store
chromadb>=0.5.0

# LLM APIs
openai>=1.50.0
google-generativeai>=0.8.0

# Arxiv Integration
arxiv>=2.1.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
tiktoken>=0.7.0

# Web UI
streamlit>=1.38.0

# Agent orchestration
langgraph>=0.2.0
